src/day4/requirements.txt
crewai>=0.100.0
crewai-tools>=0.17.0
psycopg2-binary>=2.9.0
qdrant-client>=1.12.0
fastembed>=0.4.0
python-dotenv>=1.0.0
chainlit>=2.0.0
deepeval>=2.0.0
langfuse>=2.50.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.3
httpx>=0.27.0
